refactor code from master branch
