+++ harcownik-backend/src/api/helper.py
@@ -1,0 +1,9 @@
+from src.db.db import SessionLocal
+from typing import Generator
+
+def get_db() -> Generator:
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
