+++ harcownik-backend/src/models/group.py
@@ -1,0 +1,14 @@
+from sqlalchemy import DateTime, Column, ForeignKey, Integer, String
+from sqlalchemy.orm import relationship
+from sqlalchemy.sql import func
+from src.db.db import Base
+
+class Group(Base):
+    __tablename__ = "group"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String)
+    number = Column(Integer)
+    szczep = Column(String)
+    city = Column(String)
+    time_created = Column(DateTime(timezone=True), server_default=func.now())
+    time_updated = Column(DateTime(timezone=True), onupdate=func.now())
