+++ harcownik-backend/src/schemas/__init__.py
@@ -1,0 +1,5 @@
+from .user import User, UserBase, CreateUser
+from .badge import Badge, BadgeBase, CreateBadge
+from .group import Group, GroupBase, CreateGroup
+from .badge_report import BadgeReport, BadgeReportBase, CreateBadgeReport
+from .level_report import LevelReport, LevelReportBase, CreateLevelReport
