+++ harcownik-backend/src/schemas/user.py
@@ -1,0 +1,21 @@
+from pydantic import BaseModel
+
+class UserBase(BaseModel):
+    first_name: str
+    last_name: str
+    email: str
+
+class CreateUser(UserBase):
+    level: str
+    function: str
+    group_id: int
+    badge_id: int
+
+class User(UserBase):
+    level: str
+    function: str
+    group_id: int
+    badge_id: int
+
+    class Config:
+        orm_mode = True
